pyproject.toml
[project]
name = "docling-test"
version = "0.1.0"
description = "Testing docling with Ollama and Tesseract OCR"
authors = [{ name = "Gustavo Santos", email = "gustavosisa@outlook.com" }]
dependencies = [
    "docling>=2.14.0",
    "tesserocr>=2.7.1",
    "ollama>=0.4.4",
    "mdformat>=0.7.21",
    "beautifulsoup4>=4.12.3",
    "fastapi>=0.116.0",
    "uvicorn>=0.35.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "galileo>=1.7.0",
    "markdownify>=1.1.0",
    "python-multipart>=0.0.20",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-ollama>=0.3.3",
    "openinference-instrumentation-langchain>=0.1.46",
    "opentelemetry-sdk>=1.34.1",
    "opentelemetry-exporter-otlp>=1.34.1",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-proto>=1.12.0",
]
readme = "README.md"
requires-python = ">= 3.10"
optional-dependencies = { dev = ["ruff>=0.8.4", "mypy>=1.14.0"] }
